refactor(backend): extract updateExploreWithDateZoom into a shared util
Move the date-zoom explore-rewrite out of ProjectService into a standalone
util (packages/backend/src/utils/QueryBuilder/dateZoom.ts) so both the sync
compile path and AsyncQueryService.getMetricQueryFields share one
implementation without a service dependency cycle. Verbatim move, no
behaviour change.
